Farebox and Administrator working...

+++ src/process/interProcessData/FareboxRequest.h
@@ -1,13 +1,9 @@
 #ifndef FAREBOXREQUEST_H
 #define FAREBOXREQUEST_H
 
-
 typedef struct tFareboxRequest {
-
 	int id;
-	unsigned long tax;
-
+	double tax;
 } FareboxRequest;
-
 
 #endif
